Make Quit implementation platform-specific
Move the terminateProc helper code to the actual package code, replacing
the app.Quit function - as we also use this in the signal handler that
we install.
