+++ cmd/adwatchd/commands/main.go
@@ -7,9 +7,7 @@ import (
 	"path/filepath"
 	"strings"
 	"sync"
-	"syscall"
 
-	"github.com/kardianos/service"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"github.com/ubuntu/adsys/internal/cmdhandler"
@@ -85,7 +83,7 @@ func New(opts ...option) *App {
 				oldVerbose := a.config.Verbose
 				oldDirs := a.config.Dirs
 				a.config = newConfig
-				// TODO: check why Verbose does not update properly
+
 				if oldVerbose != a.config.Verbose {
 					config.SetVerboseMode(a.config.Verbose)
 				}
@@ -220,25 +218,6 @@ func (a *App) Reset() {
 	a.ready = make(chan struct{})
 }
 
-// Quit gracefully exits the app. Shouldn't be in general necessary apart for
-// integration tests where we might need to close the app manually.
-func (a *App) Quit(sig syscall.Signal) error {
-	a.waitReady()
-	if !service.Interactive() {
-		return fmt.Errorf(i18n.G("not running in interactive mode"))
-	}
-
-	// The service package is responsible for handling the service stop. It
-	// registers a signal handler and to trigger it we just have to send the
-	// signal ourselves and not bother with actual cleanup.
-	p, err := os.FindProcess(os.Getpid())
-	if err != nil {
-		return err
-	}
-
-	return p.Signal(sig)
-}
-
 // WithServiceName allows setting a custom name for the daemon. Shouldn't be in
 // general necessary apart for integration tests where it helps with parallel
 // execution.
@@ -248,8 +227,8 @@ func WithServiceName(name string) func(o *options) {
 	}
 }
 
-// waitReady signals when the daemon is ready
+// WaitReady signals when the daemon is ready
 // Note: we need to use a pointer to not copy the App object before the daemon is ready, and thus, creates a data race.
-func (a *App) waitReady() {
+func (a *App) WaitReady() {
 	<-a.ready
 }
